pyproject.toml
[tool.poetry]
name = "hexlet_gendiff"
version = "0.2.0"
description = "JSON diffing tool"
authors = ["Hexlet team <team@hexlet.io>"]
license = "MIT"
readme = "README.md"
homepage = "https://hexlet.io"
packages = [
  { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
